Move ObsTable and XTable to new components
